+++ CHANGES.rst
@@ -28,10 +28,17 @@ Unreleased
     wikipage submissions.
  * **[CHANGE]** Increase the dependency of ``update_checker`` to 0.10 or later
    to prevent ImportWarnings (issue 291).
+ * **[CHANGE]** :meth:`~.Subreddit.get_banned` now takes a `user_only` argument
+     (default: ``True``). When the value is explicitly passed as ``False`` the
+     return value is not a generator of ``Redditor`` objects, but a generator
+     of dictionaries whose ``name`` key corresponds to the ``Redditor`` object
+     and whose ban-note is at key ``note``.
  * **[FEATURE]** Enable gathering of duplicate submissions for a Submission
    object (issue 290).
  * **[FEATURE]** Add :meth:`~praw.__init__.LoggedInReddior.delete`.
-
+ * **[FEATURE]** ``Redditor`` objects returned via a user listing now have a
+     ``userlist_data`` attribute which contains addition information provided
+     by reddit's API.
 
 PRAW 2.1.15
 -----------